CA-220044: xcp-rrdd-plugins: don't capture stdout
These plugins produce a lot of logging on stdout when not daemonized.  This is
captured by systemd and results in a lot of spam in daemon.log.

This spam wasn't present when run daemonized, as was the case when run by the
SYSV init script.

Disabling this does not affect the logging done by the plugins via syslog (for
which info-level and above logging normally ends up in
/var/log/xcp-rrdd-plugins.log).

Signed-off-by: Jonathan Davies <jonathan.davies@citrix.com>